requirements.txt
asyncio
discord.py
PyNaCl
python-dotenv
pytube
openai